# Conflicts:
#	crates/re_viewer/src/app.rs
#	crates/re_viewer/src/store_hub.rs

+++ rerun_cpp/src/rerun/datatypes/mat4x4.cpp
@@ -39,11 +39,13 @@ namespace rerun {
 
             auto value_builder = static_cast<arrow::FloatBuilder *>(builder->value_builder());
 
-            ARROW_RETURN_NOT_OK(builder->AppendValues(num_elements));
+            ARROW_RETURN_NOT_OK(builder->AppendValues(static_cast<int64_t>(num_elements)));
             static_assert(sizeof(elements[0].coeffs) == sizeof(elements[0]));
-            ARROW_RETURN_NOT_OK(
-                value_builder->AppendValues(elements[0].coeffs, num_elements * 16, nullptr)
-            );
+            ARROW_RETURN_NOT_OK(value_builder->AppendValues(
+                elements[0].coeffs,
+                static_cast<int64_t>(num_elements * 16),
+                nullptr
+            ));
 
             return arrow::Status::OK();
         }
